moving convolution code to its own file
